+++ quartz/src/androidMain/kotlin/com/vitorpamplona/quartz/nip72ModCommunities/approval/tags/ApprovedAddressTag.kt
@@ -36,7 +36,7 @@ class ApprovedAddressTag(
     val address: Address,
     val relayHint: NormalizedRelayUrl? = null,
 ) {
-    fun countMemory(): Long = 2 * pointerSizeInBytes + address.countMemory() + (relayHint?.url?.bytesUsedInMemory() ?: 0)
+    fun countMemory(): Int = 2 * pointerSizeInBytes + address.countMemory() + (relayHint?.url?.bytesUsedInMemory() ?: 0)
 
     fun toTag() = Address.assemble(address.kind, address.pubKeyHex, address.dTag)
